Flytter mapTilRegisterAktiviteterObjektMedLabel til felles util for registerAktivitet for å kunne gjenbruke fra læremidler
