+++ src/frontend/barnetilsyn/steg/3-aktivitet/utils.ts
@@ -1,22 +1,6 @@
 import { AnnenAktivitetType } from '../../../typer/aktivitet';
-import { RegisterAktivitetMedLabel, RegisterAktivitet } from '../../../typer/registerAktivitet';
+import { RegisterAktivitetMedLabel } from '../../../typer/registerAktivitet';
 import { EnumFelt, EnumFlereValgFelt } from '../../../typer/skjema';
-import { tilTekstligDato } from '../../../utils/dato';
-
-export const mapTilRegisterAktiviteterObjektMedLabel = (
-    registerAktiviteter: RegisterAktivitet[]
-): Record<string, RegisterAktivitetMedLabel> => {
-    return registerAktiviteter.reduce(
-        (acc, curr) => ({
-            ...acc,
-            [curr.id]: {
-                ...curr,
-                label: `${curr.typeNavn}: ${tilTekstligDato(curr.fom)} - ${curr.tom ? tilTekstligDato(curr.tom) : ''}`,
-            },
-        }),
-        {} as Record<string, RegisterAktivitetMedLabel>
-    );
-};
 
 export const skalTaStillingTilAnnenAktivitet = (
     valgteAktiviteter: EnumFlereValgFelt<string> | undefined
